Tometo-Typing-App
The "Tomato App" is a multifunctional text processing tool designed to enhance text manipulation capabilities. Users can easily convert text to uppercase or lowercase, making it adaptable to specific formatting needs. It also offers a convenient feature to remove extra spaces from paragraphs, streamlining text for cleaner presentations or code formatting.

Furthermore, the app simplifies the process of copying text with a user-friendly interface, ensuring smooth and efficient text transfer. The inclusion of dark and light modes caters to users' preferences, providing a visually comfortable environment for extended usage, whether in low-light conditions or during the day.

Overall, the "Tomato App" is a versatile tool that simplifies text editing tasks, streamlining the process of text transformation and enhancing the reading experience with its customizable visual modes. Whether you're working on documents, coding, or simply want to improve the readability of text, this app is a valuable asset for various text-related projects and activities